api/pyproject.toml
[project]
name = "api"
version = "0.1.1"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.11",
    "pydantic-settings>=2.8.1",
    "pydantic>=2.10.6",
    "pyjwt[crypto]>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
    "email-validator>=2.2.0",
]


[tool.mypy]
python_version = "3.12"
warn_return_any = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
